+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - [INTERNAL] Completely revamp comment parsing by switching off the internal lexer state from `ripper`. This should drastically increase accuracy of comment parsing in general, and set us up for success in the future.
 - Allow comments to be attached to `CHAR` nodes.
 - [INTERNAL] Disallow comments from being attached to `args_new` nodes.
+- [INTERNAL] Track start and end lines so we can better insert block comments.
 
 ## [0.1.2] - 2019-02-05
 ### Changed